+++ storage/innobase/include/trx0sys.h
@@ -704,9 +704,10 @@ public:
     because it may change even before this method returns.
   */
 
-  int32_t size()
+  uint32_t size()
   {
-    return my_atomic_load32_explicit(&hash.count, MY_MEMORY_ORDER_RELAXED);
+    return uint32_t(my_atomic_load32_explicit(&hash.count,
+					      MY_MEMORY_ORDER_RELAXED));
   }
